lab 2 part 2.0

+++ projects/lpc40xx_freertos/l5_application/gpio_lab.h
@@ -5,8 +5,8 @@
 #include <stdint.h>
 
 typedef struct {
-  uint8_t port;
-  uint8_t pin;
+  int port;
+  int pin;
 } port_pin_s;
 
 /// Should alter the hardware registers to set the pin as input
